Move iOS Motion check to initial title page and set globally
