+++ client/wfnews-war/src/main/angular/src/app/components/report-of-fire/reportOfFireModel.ts
@@ -26,4 +26,6 @@ export class ReportOfFire {
   public currentLocation: Array<number> = [0, 0];
   public compassHeading: number = 0;
   public headingDetectionActive: boolean;
+  public iosGranted: boolean;
+  public androidGranted: boolean;
 }
