+++ client/wfnews-war/src/main/angular/src/app/components/report-of-fire/title-page/rof-title-page.component.ts
@@ -1,9 +1,15 @@
-import { Component, ChangeDetectionStrategy } from "@angular/core";
+import { Component, ChangeDetectionStrategy, OnInit } from "@angular/core";
 import { RoFPage } from "../rofPage";
 import { ReportOfFire } from "../reportOfFireModel";
 import { MatDialog } from "@angular/material/dialog";
 import { DialogLocationComponent } from "@app/components/report-of-fire/dialog-location/dialog-location.component";
 import { CommonUtilityService } from "@app/services/common-utility.service";
+import { LocationServicesDialogComponent } from "../compass-page/location-services-dialog/location-services-dialog.component";
+import { equalsIgnoreCase } from "@app/utils";
+
+interface DeviceOrientationEventiOS extends DeviceOrientationEvent {
+  requestPermission?: () => Promise<'granted' | 'denied'>;
+}
 
 @Component({
   selector: 'rof-title-page',
@@ -11,7 +17,7 @@ import { CommonUtilityService } from "@app/services/common-utility.service";
   styleUrls: ['./rof-title-page.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class RoFTitlePage extends RoFPage {
+export class RoFTitlePage extends RoFPage implements OnInit {
   public imageUrl: string
   public closeButton: boolean
   public messages: any;
@@ -35,12 +41,16 @@ export class RoFTitlePage extends RoFPage {
     this.offLine = !window.navigator.onLine;
   }
 
+  ngOnInit() {
+    this.getOrientation();
+  }
+
   openCallPage () {
     // not yet implemented
   }
 
 
-  triggerLocationServiceCheck(){
+ triggerLocationServiceCheck (){
     this.commonUtilityService.checkLocationServiceStatus().then((enabled) => {
       if (!enabled) {
         let dialogRef = this.dialog.open(DialogLocationComponent, {
@@ -52,4 +62,36 @@ export class RoFTitlePage extends RoFPage {
       }
     });
   }
+
+  async getOrientation() {
+    try{
+      const requestPermission = (DeviceOrientationEvent as unknown as DeviceOrientationEventiOS).requestPermission;
+      const iOS = typeof requestPermission === 'function';
+      if (iOS) {
+      const response = await requestPermission();
+          if (equalsIgnoreCase(response, "granted")) {
+            this.reportOfFire.iosGranted = true;
+          } else {
+              this.dialog.open(LocationServicesDialogComponent, {
+              width: '350px',
+              data: {
+                message: "Location services are required"
+              }
+            });
+          }
+        } else {
+          this.reportOfFire.androidGranted = true;
+       }
+       } catch (err) {
+        this.dialog.open(LocationServicesDialogComponent, {
+          width: '350px',
+          data: {
+              message: "Location services are not supported"
+          }
+        }); 
+    } 
+  
+  }
+
+  
 }
